requirements.txt
# Core dependencies
streamlit==1.31.0
python-dotenv==1.0.0

# LLM and LangChain
langchain==0.1.6
langchain-community==0.0.20
langchain-openai==0.0.5
openai==1.12.0

# Graph database
neo4j==5.16.0

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
pandas==2.2.0

# NLP and entity extraction
spacy==3.7.2

# Visualization
pyvis==0.3.2
plotly==5.18.0
networkx==3.2.1

# Utilities
tqdm==4.66.1
pydantic==2.6.0
requests==2.31.0
